Revert "fixed compiler warnings"
This reverts commit f2bb2d9f2e6f56f0c9990a3fafc8f07eda952c3b.

+++ MecabCocoa/dictionary.cpp
@@ -244,7 +244,7 @@ bool Dictionary::compile(const Param &param,
       token->rcAttr = rid;
       token->posid  = pid;
       token->wcost = cost;
-      token->feature = static_cast<uint>(offset);
+      token->feature = offset;
       token->compound = 0;
       dic.push_back(std::make_pair<std::string, Token*>(w, token));
 
@@ -274,7 +274,7 @@ bool Dictionary::compile(const Param &param,
     if (i != 0 && prev != dic[i].first) {
       str.push_back(dic[idx].first.c_str());
       len.push_back(dic[idx].first.size());
-      val.push_back(static_cast<int>(bsize +(idx << 8)));
+      val.push_back(bsize +(idx << 8));
       bsize = 1;
       idx = i;
     } else {
@@ -284,7 +284,7 @@ bool Dictionary::compile(const Param &param,
   }
   str.push_back(dic[idx].first.c_str());
   len.push_back(dic[idx].first.size());
-  val.push_back(static_cast<int>(bsize +(idx << 8)));
+  val.push_back(bsize +(idx << 8));
 
   CHECK_DIE(str.size() == len.size());
   CHECK_DIE(str.size() == val.size());
@@ -310,11 +310,11 @@ bool Dictionary::compile(const Param &param,
   }
 
   unsigned int dummy = 0;
-  unsigned long lsize = matrix.left_size();
-  unsigned long rsize = matrix.right_size();
-  unsigned long dsize = da.unit_size() * da.size();
-  unsigned long tsize = tbuf.size();
-  unsigned long fsize = fbuf.size();
+  unsigned int lsize = matrix.left_size();
+  unsigned int rsize = matrix.right_size();
+  unsigned int dsize = da.unit_size() * da.size();
+  unsigned int tsize = tbuf.size();
+  unsigned int fsize = fbuf.size();
 
   unsigned int version = DIC_VERSION;
   char charset[32];
